refactor: rename model module
